services/users/requirements.txt
black==19.3b0
flake8===3.7.8
Flask==1.1.1
Flask-Admin==1.5.3
flask-cors==3.0.8
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
isort==4.3.21
psycopg2-binary==2.8.3
pytest==5.0.1
pytest-cov==2.7.1
